add good against heroes feature

+++ web_scrap/scrap.py
@@ -5,6 +5,8 @@ from constant import CT_IMAGE_PATH, CT_IMAGE_UPDATE_TIME_THRESHOLD
 import constant
 import os
 import cv2
+from PIL import Image, ImageEnhance
+import numpy as np
 
 
 def round_df_digits(df):
@@ -37,13 +39,13 @@ def add_border_to_image(im, bordersize=5, rgb=[45, 33, 31]):
     return border
 
 
-def make_counter_hero_image(hero_image_path, counter_heros_image_path):
-    bg_image = cv2.imread(constant.COUNTER_BG_IMAGE_PATH)
+def make_hero_images(main_hero_image_path, heroes_image_path, bg_path):
+    bg_image = cv2.imread(bg_path)
     bg_image = cv2.resize(bg_image, (constant.COUNTER_BG_SHAPE[1], constant.COUNTER_BG_SHAPE[0]))
-    main_hero_image = cv2.imread(hero_image_path)
+    main_hero_image = cv2.imread(main_hero_image_path)
     X, Y = constant.COUNTER_MAIN_HERO_COORDS
     bg_image[X: X + main_hero_image.shape[0], Y: Y + main_hero_image.shape[1], :] = main_hero_image
-    for i, path in enumerate(counter_heros_image_path):
+    for i, path in enumerate(heroes_image_path):
         image = cv2.imread(path)
         image = cv2.resize(image, (constant.COUNTER_ICON_SHAPE[1], constant.COUNTER_ICON_SHAPE[0]))
         image = add_border_to_image(image)
@@ -52,7 +54,6 @@ def make_counter_hero_image(hero_image_path, counter_heros_image_path):
         y = y + ((i % constant.COUNTER_MAX_COLUMN) * image.shape[1]) +\
             ((i % constant.COUNTER_MAX_COLUMN) * constant.COUNTER_WIDTH_DIST)
         bg_image[x: x + image.shape[0], y: y + image.shape[1], :] = image
-
     return bg_image
 
 
@@ -73,12 +74,35 @@ def get_counter_hero(query):
     counter_heroes_list = list(counter_heroes)
     counter_heroes_image_path = get_icon_path(counter_heroes_list, icon_size='big')
     hero_image_path = get_icon_path([hero_name], icon_size='big')[0]
-    image = make_counter_hero_image(hero_image_path, counter_heroes_image_path)
+    image = make_hero_images(hero_image_path, counter_heroes_image_path, constant.COUNTER_BG_IMAGE_PATH)
+    cv2.imwrite(image_path, image)
+    return True, hero_name, image_path
+
+
+def get_good_against(query):
+    query = query.split()
+    hero = ' '.join(query[1:])
+    hero = hero.strip()
+    found_hero, hero_name = find_hero_name(hero)
+    if not found_hero:
+        return False, hero_name, ''
+    image_path = os.path.join(constant.GOOD_HERO_IMAGE_PATH, hero_name + '.png')
+    if not is_file_old(image_path, constant.COUNTER_HERO_UPDATE_TIME_THRESHOLD):
+        return True, hero_name, image_path
+
+    hero_info = scrap_heroes_info(hero_name)
+    good_against_info = hero_info[1]
+    good_against_heroes = good_against_info['Hero']
+    good_against_heroes_list = list(good_against_heroes)
+    good_against_heroes_image_path = get_icon_path(good_against_heroes_list, icon_size='big')
+    hero_image_path = get_icon_path([hero_name], icon_size='big')[0]
+    image = make_hero_images(hero_image_path, good_against_heroes_image_path, constant.GOOD_BG_IMAGE_PATH)
     cv2.imwrite(image_path, image)
     return True, hero_name, image_path
 
 
 if __name__ == '__main__':
-    print(get_counter_hero('anti mage'))
+    print(get_counter_hero('!good ursa'))
+    print(get_good_against('!good ursa'))
